task7

+++ controller/expense.js
@@ -8,7 +8,7 @@ exports.addExpense = (req,res,next) => {
         {
             return res.status(400).json({err:'Parameters Missing'})
         } else {
-            Expense.create({amount,description,category})
+            Expense.create({amount,description,category,userId:req.user.id})
             .then(expense=>{
                 res.status(201).json({message:'Expense added',success:true,expense})
 
@@ -22,14 +22,19 @@ exports.addExpense = (req,res,next) => {
 
 
 exports.showExpense = (req,res,next)=>{
-    Expense.findAll()
+    Expense.findAll({where:{userId: req.user.id}})
     .then(expenses=>{
+        console.log(expenses);
         return res.status(200).json({expenses,success:true})
     })
 }
 exports.deleteexpense = (req, res) => {
     const expenseid = req.params.expenseid;
-    Expense.destroy({where: { id: expenseid }}).then(() => {
+    Expense.destroy({where: { id: expenseid,userId:req.user.id }}).then((noOfrows) => {
+if(noOfrows===0){
+    return res.status(404).json({success:false,message:'This Expense Does not belong to this user'})
+}
+
         return res.status(204).json({ success: true, message: "Deleted Successfuly"})
     }).catch(err => {
         console.log(err);
